Final changes for pre and post stages
