+++ lib/base_extension.rb
@@ -1,22 +1,10 @@
 # frozen_string_literal: true
 
 require_relative './extension_manager.rb'
+require_relative './utils/paths.rb'
 
 module Toolchain
   require 'zlib'
-  ##
-  # content_path
-  # Returns path to content directory +content_dir_path+.
-  #
-  def self.content_path(path = nil)
-    content_dir_path = '..'
-    content_dir_path = ENV['GITHUB_WORKSPACE'] \
-      if ENV.key?('TOOLCHAIN_TEST') || ENV.key?('GITHUB_ACTIONS')
-    content_dir_path = ENV['CONTENT_PATH'] if ENV.key?('CONTENT_PATH')
-    # For Unit testing:
-    content_dir_path = path unless path.nil?
-    return content_dir_path
-  end
 
   # Location describes a source location, made up of +filename+ and +lineno+.
   # Params:
